Extract transforms and utils
Preparation work for #3: separated existing (classic component and template) transforms and utils into separate modules, so we can later add a native class transform easier.
