+++ lib/__tests__/find-properties.js
@@ -6,7 +6,7 @@ const {
   findAttributeBindings,
   findClassNames,
   findClassNameBindings,
-} = require('../transform');
+} = require('../utils');
 
 describe('findTagName()', () => {
   const TESTS = [
