+++ lib/__tests__/transform.js
@@ -106,7 +106,7 @@ test('throws if `Component.extend({ ... })` is not found', () => {
   `;
 
   expect(() => transform(source, '')).toThrowErrorMatchingInlineSnapshot(
-    `"Could not find \`export default Component.extend({ ... });\`"`
+    `"Unsupported component type. Only classic components (\`Component.extend({ ... }\`) are supported currently."`
   );
 });
